+++ controllers/routers/index.js
@@ -1,9 +1,10 @@
-const { blog, editor } = require('./blog');
 const { auth } = require('./auth');
+const { blog } = require('./blog');
 const { home } = require('./home');
+const { init } = require('./init');
 const { page } = require('./page');
 const { site } = require('./site');
 
 
 // exports
-module.exports = { auth, blog, editor, home, page, site };
+module.exports = { auth, blog, home, init, page, site };