+++ models/configs.js
@@ -4,6 +4,7 @@ const mongoose = require('mongoose');
 
 // schema
 const configsSchema  = new mongoose.Schema({
+  initialized        : { type: Boolean, default: false },
   active             : { type: Boolean, default: false, unique: true },
   title              : { type: String, default: 'New Website' },
   description        : { type: String, default: 'n/a' },
@@ -23,17 +24,16 @@ const configsSchema  = new mongoose.Schema({
 
 
 // static methods
-configsSchema.static('initialize', async function configsSchema_initialize(next = () => {}) {
-  process.env.$WEBSITE_CONFIGS = JSON.stringify(await this.findOne({ active: true }));
-  if (process.env.$WEBSITE_CONFIGS === 'null') {
-    process.env.$WEBSITE_CONFIGS = JSON.stringify(await this.create({ active: true }));
-  } return next();
+configsSchema.static('initConfig', async function configsSchema_initialize(app, cb) {
+  const config = await this.findOne({ active: true });
+  app.set('APP_CONFIG', config ? (await config.toObject()) : (await this.create({ active: true })).toObject());
+  if (typeof cb === 'function') return cb();
 });
 
 
-configsSchema.static('updateConfigs', async function configsSchema_updateConfigs(doc, next = () => {}) {
-  process.env.$WEBSITE_CONFIGS = JSON.stringify(await this.findOneAndUpdate({ active: true }, doc, { new: true }));
-  return next();
+configsSchema.static('updateConfig', async function configsSchema_updateConfig(app, doc, cb) {                          // todo: update lang (process.env)
+  app.set('APP_CONFIG', (await this.findOneAndUpdate({ active: true }, doc, { new: true })).toObject());
+  if (typeof cb === 'function') return cb();
 });